requirements.txt
fastapi==0.115.0
uvicorn==0.30.6

pandas
numpy
scikit-learn
joblib

qiskit
qiskit-machine-learning

sqlalchemy
psycopg2-binary
python-dotenv

passlib[bcrypt]
python-jose
python-multipart

pydantic
typing-extensions